Make all unit tests pass
