+++ src/SpecFirst.Core/Serialization/NullDataSerializer.cs
@@ -1,0 +1,14 @@
+﻿namespace SpecFirst.Core.Serialization
+{
+    using System.Diagnostics;
+
+    public class NullDataSerializer
+    {
+        public string Serialize(object data)
+        {
+            Debug.Assert(data == null);
+
+            return $"{"null".ToLowerInvariant()}";
+        }
+    }
+}
