+++ tests/SpecFirst.Core.Specs/SpecFirst.Core.Specs.csproj
@@ -28,7 +28,7 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
 		<PackageReference Include="coverlet.collector" Version="1.2.0" />
-		<PackageReference Include="SpecFirst.xUnit" Version="1.0.0-rc.32" />
+		<PackageReference Include="SpecFirst.xUnit" Version="1.1.0-rc.9" />
 	</ItemGroup>
 
   <ItemGroup>
